Mount app as a node module
This allows me to require() modules from the app without awkward
relative paths that have a various number of '../'s.

For this, I had to give up on requiring CoffeeScript modules without the
'.coffee' extension, but that seems like a good trade now.

Because:

Mounting the app as a node module, seems to require the
Browserify-specific package.json in the mounted directory
(./app/package.json in this case) so that Browserify knows that the
imported module is not raw JS.

But then, I have to get the transform out of the server’s
browserify-middleware configuration, to prevent .coffee files from being
coffeeified twice.

But then, this also requires a Broserify-specific package.json for the
other entry too: ./test/package.json.