Tooltip: a failing test
I want a way to show a warning message to guide the user when the client
list contains innapropriate characters.

I want to be able to set the text when I want, be able to show it at
specific times, and hide it. Also let the user close it.

I couldn’t google any.